CU-51trg1 move file/database entitys
