+++ src/main/java/org/elastos/hive/database/DocCount.java
@@ -1,0 +1,9 @@
+package org.elastos.hive.database;
+
+public class DocCount {
+	private long count;
+
+	public long getCount() {
+		return count;
+	}
+}
