+++ src/main/java/org/elastos/hive/files/FilesList.java
@@ -1,10 +1,10 @@
-package org.elastos.hive.network.model;
+package org.elastos.hive.files;
 
 import java.util.List;
 
 import org.elastos.hive.files.FileInfo;
 
-public class FilesResponse {
+public class FilesList {
 	private List<FileInfo> file_info_list;
 
 	public List<FileInfo> getFiles() {
