--- src/main/java/org/elastos/hive/network/model/CountDocResponse.java
@@ -1,9 +1,0 @@
-package org.elastos.hive.network.model;
-
-public class CountDocResponse {
-	private long count;
-
-	public long getCount() {
-		return count;
-	}
-}
