--- src/main/java/org/elastos/hive/network/model/UploadResponse.java
@@ -1,9 +1,0 @@
-package org.elastos.hive.network.model;
-
-public class UploadResponse {
-	private String upload_file_url;
-
-	public String getUpload_file_url() {
-		return upload_file_url;
-	}
-}
